Refactor parts into their own classes
Big refactor that changes question structure from 25 day classes each
containing a Part1 and Part2, to 25 day namespaces each of which
*should* contain a Part1 and Part2 class that implement IAocAnswer,
requiring them to contain a long Run().

Done primarily to separate concerns. There is no reason that an answer
part should be able to see methods and classes that it has no need of,
so they now become private in each part's class. Shared methods
and classes now reside in the static Common class in the part's
namespace.

No changes to the test suite are required.